requirements.txt
pandas>=2.0
numpy>=1.24
ccxt>=4.0
pyyaml>=6.0
pytest>=7.0
